Add video viewer HTML with custom controls
This file is the entire app (besides for manifest.json and sw.js, to make it useable as a PWA (progressive web app)) as a standalone .html file.  It implements a video viewer with custom playback speed controls, and forward/backward 1 frame controls through tapping the right/left side of the screen.